+++ docs/releasenotes.md
@@ -12,6 +12,7 @@ Where:
 ## Release 6.4.2 - not yet released
 
 * Fix [#210](https://lab.civicrm.org/extensions/stripe/-/issues/210): If there are multiple reCaptcha on the page check and validate the one on the Stripe billing form only.
+* Update implementation for cancel subscription. Option to notify (default Yes) is now available on the backend cancel subscription form.
 
 ## Release 6.4.1
 **This release REQUIRES that you upgrade mjwshared to 0.8.**